# apps/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ProSlides API
  version: 0.1.0
  description: Contract-first API for the Go modular monolith.
servers:
  - url: /
paths:
...
  /api/v1/presentations/{presentationId}:
    get:
      operationId: getPresentation
      summary: Return one presentation and its ordered slides for its owner.
      description: Requires the current authenticated session. Returns 404 when the presentation does not exist or is not owned by the current account, so identifiers are not disclosed.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
      responses:
        '200':
          description: The presentation with slides ordered by ascending position.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Presentation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updatePresentation
      summary: Atomically update an owned presentation's title or visual settings.
      description: Settings are merged by key. Clients should send the last observed revision in If-Match; a stale revision returns 409 without overwriting newer edits.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/RevisionMatch'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdatePresentationRequest' } } }
      responses:
        '200': { description: Updated presentation., content: { application/json: { schema: { $ref: '#/components/schemas/Presentation' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/EditConflict' }
    delete:
      operationId: deletePresentation
      summary: Permanently delete an owned presentation and its dependent data.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Presentation deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: The presentation has a non-ended live session and cannot be deleted. }
  /api/v1/presentations/{presentationId}/access-code:
    put:
      operationId: setPresentationAccessCode
      summary: Set the public access code for an owned presentation.
      description: Normalizes the code to uppercase and updates any non-ended live session for this presentation atomically, immediately invalidating its previous public link. Codes are globally unique without regard to case.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SetAccessCodeRequest' } } }
      responses:
        '200': { description: Access code saved., content: { application/json: { schema: { $ref: '#/components/schemas/AccessCodeResult' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: The access code is already assigned to another presentation. }
...
  /api/v1/presentations:
    get:
      operationId: listPresentations
      summary: List the current owner's presentations in stable recent-first order.
      responses:
        '200':
          description: Bounded owner presentation list.
          content: { application/json: { schema: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/PresentationSummary' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createPresentation
      summary: Create an empty presentation owned by the current account.
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePresentationRequest' }
      responses:
        '201':
          description: Presentation created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Presentation' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
  /api/v1/presentations/{presentationId}/duplicate:
    post:
      operationId: duplicatePresentation
      summary: Atomically duplicate an owned presentation and all ordered slides.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreatePresentationRequest' } } }
      responses:
        '201': { description: Duplicated presentation., content: { application/json: { schema: { $ref: '#/components/schemas/Presentation' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/presentations/{presentationId}/results:
    delete:
      operationId: deletePresentationResults
      summary: Delete all live sessions and results for an owned presentation while preserving its content.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Results deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Results cannot be reset while the presentation has a non-ended live session. }
  /api/v1/presentations/{presentationId}/slides:
    post:
      operationId: createSlide
      summary: Add a slide to an owned presentation.
      description: Requires an authenticated owner and a CSRF token. Inserting at an occupied zero-based position atomically shifts following slides. Returns 404 for absent or non-owned presentations.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/RevisionMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSlideRequest' }
      responses:
        '201':
          description: Slide created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Slide' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/EditConflict' }
  /api/v1/presentations/{presentationId}/slides/reorder:
    post:
      operationId: reorderSlides
      summary: Atomically replace the complete slide order for an owned presentation.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/RevisionMatch'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReorderSlidesRequest' } } }
      responses:
        '204': { description: Slide positions updated. }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/EditConflict' }
  /api/v1/presentations/{presentationId}/slides/{slideId}:
    put:
      operationId: replaceSlide
      summary: Replace the editable definition of one owned slide.
      description: Moving to an occupied zero-based position atomically compacts the other positions. A slide with durable answers may be moved, but its question definition cannot change until presentation results are reset.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - { name: slideId, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/RevisionMatch'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateSlideRequest' } } }
      responses:
        '200': { description: Replaced slide., content: { application/json: { schema: { $ref: '#/components/schemas/Slide' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: The requested edit would change a slide with durable live answers. }
    delete:
      operationId: deleteSlide
      summary: Delete one owned slide and compact following positions.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - { name: slideId, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/RevisionMatch'
      responses:
        '204': { description: Slide deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: The slide has durable live answers; reset presentation results before deleting it. }
...
components:
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
...
    PresentationId:
      name: presentationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RevisionMatch:
      name: If-Match
      in: header
      required: false
      description: Last observed positive integer revision. When supplied, a stale value returns 409 rather than overwriting a newer edit.
      schema: { type: integer, format: int64, minimum: 1 }
  schemas:
...
    Presentation:
      type: object
      required: [id, revision, title, access_code, settings, created_at, updated_at, slides]
      properties:
        id: { type: string, format: uuid }
        revision: { type: integer, format: int64, minimum: 1 }
        title: { type: string, minLength: 1, maxLength: 500 }
        access_code: { type: string, pattern: '^[A-Z0-9]{5,12}$', nullable: true }
        settings: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        slides:
          type: array
          items: { $ref: '#/components/schemas/Slide' }
    PresentationSummary:
      type: object
      required: [id, revision, title, access_code, settings, slide_count, participant_count, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        revision: { type: integer, format: int64, minimum: 1 }
        title: { type: string }
        access_code: { type: string, pattern: '^[A-Z0-9]{5,12}$', nullable: true }
        settings: { type: object, additionalProperties: true }
        slide_count: { type: integer, minimum: 0 }
        participant_count: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PresentationSettingsInput:
      type: object
      additionalProperties: true
      properties:
        background_color: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
        background_image_url:
          type: string
          maxLength: 4096
          description: Legacy remote URL or stable first-party immutable media content path.
        background_image_asset_id:
          type: string
          maxLength: 36
          description: Optional immutable Media Asset UUID. Empty string clears the asset reference.
        background_image_focal_x:
          type: number
          format: double
          minimum: 0
          maximum: 1
          default: 0.5
          description: Normalized horizontal focal coordinate used when the background covers the Stage.
        background_image_focal_y:
          type: number
          format: double
          minimum: 0
          maximum: 1
          default: 0.5
          description: Normalized vertical focal coordinate used when the background covers the Stage.
        music_url:
          type: string
          maxLength: 4096
          pattern: '^\s*(?:https?://[^/\s?#]+(?:[/?#]\S*)?)?\s*$'
        text_color: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
        accent_color: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
        visualization_palette:
          type: array
          minItems: 3
          maxItems: 8
          items: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
    CreatePresentationRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        settings: { $ref: '#/components/schemas/PresentationSettingsInput' }
    UpdatePresentationRequest:
      type: object
      minProperties: 1
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        settings: { $ref: '#/components/schemas/PresentationSettingsInput' }
    SetAccessCodeRequest:
      type: object
      required: [access_code]
      additionalProperties: false
      properties:
        access_code: { type: string, pattern: '^[A-Za-z0-9]{5,12}$' }
    AccessCodeResult:
      type: object
      required: [access_code]
      additionalProperties: false
      properties:
        access_code: { type: string, pattern: '^[A-Z0-9]{5,12}$' }
    CreateSlideRequest:
      type: object
      required: [position, kind, content]
      additionalProperties: false
      properties:
        position: { type: integer, minimum: 0 }
        kind: { type: string, enum: [activity, content] }
        content: { $ref: '#/components/schemas/EditableSlideContent' }
    ReorderSlidesRequest:
      type: object
      required: [slide_ids]
      properties:
        slide_ids:
          type: array
          maxItems: 1000
          uniqueItems: true
          items: { type: string, format: uuid }
    EditableSlideContent:
      description: Content is validated against the accompanying authoring item kind. Unknown fields are rejected by the API.
      anyOf:
        - $ref: '#/components/schemas/ActivityItemDefinition'
        - $ref: '#/components/schemas/ContentSlideContent'
    ActivityItemDefinition:
      description: Versioned audience Activity definition. Concrete response/evaluation/result policies are selected by activity_kind; no generic capability bag is persisted.
      oneOf:
        - $ref: '#/components/schemas/ChoiceActivityDefinition'
        - $ref: '#/components/schemas/TextActivityDefinition'
    ChoiceActivityDefinition:
      type: object
      additionalProperties: false
      required: [schema_version, activity_kind, prompt, response, evaluation, scoring, timing, results]
      properties:
        schema_version: { type: integer, enum: [1] }
        activity_kind: { type: string, enum: [choice] }
        prompt: { $ref: '#/components/schemas/ActivityPrompt' }
        response: { $ref: '#/components/schemas/ChoiceResponsePolicy' }
        evaluation: { $ref: '#/components/schemas/ChoiceEvaluationPolicy' }
        scoring: { $ref: '#/components/schemas/ChoiceScoringPolicy' }
        timing: { $ref: '#/components/schemas/ActivityTimingPolicy' }
        results: { $ref: '#/components/schemas/ActivityResultPolicy' }
    TextActivityDefinition:
      type: object
      additionalProperties: false
      required: [schema_version, activity_kind, prompt, response, evaluation, scoring, timing, results]
      properties:
        schema_version: { type: integer, enum: [1] }
        activity_kind: { type: string, enum: [text] }
        prompt: { $ref: '#/components/schemas/ActivityPrompt' }
        response: { $ref: '#/components/schemas/TextResponsePolicy' }
        evaluation: { $ref: '#/components/schemas/TextEvaluationPolicy' }
        scoring: { $ref: '#/components/schemas/TextScoringPolicy' }
        timing: { $ref: '#/components/schemas/ActivityTimingPolicy' }
        results: { $ref: '#/components/schemas/TextResultPolicy' }
    TextResponsePolicy:
      description: Legacy Word Clouds use word limits; entry-based Word Clouds preserve short phrases as independent entries.
      oneOf:
        - $ref: '#/components/schemas/LegacyTextResponsePolicy'
        - $ref: '#/components/schemas/EntryTextResponsePolicy'
...
    ActivityPrompt:
      type: object
      additionalProperties: false
      required: [title, text, image_url]
      properties:
        title: { type: string, maxLength: 80 }
        text: { type: string, minLength: 1, maxLength: 180 }
        image_url: { type: string, maxLength: 4096 }
        image_asset_id: { type: string, format: uuid }
        image_width: { type: integer, minimum: 1, maximum: 8192 }
        image_height: { type: integer, minimum: 1, maximum: 8192 }
        image_alt_text: { type: string, maxLength: 300 }
        image_focal_x: { type: number, format: double, minimum: 0, maximum: 1 }
        image_focal_y: { type: number, format: double, minimum: 0, maximum: 1 }
    ChoiceResponsePolicy:
      type: object
      additionalProperties: false
      required: [selection, options]
      properties:
        selection: { type: string, enum: [single, multiple] }
        options:
          type: array
          minItems: 2
          maxItems: 12
          description: Presenter-paced live density limit. Correctness-evaluated Choice activities allow at most 8 options; unscored Poll activities allow at most 12.
          items: { $ref: '#/components/schemas/ChoiceOptionDefinition' }
...
    ChoiceEvaluationPolicy:
      type: object
      additionalProperties: false
      required: [mode, correct_option_ids]
      properties:
        mode: { type: string, enum: [none, correctness] }
        correct_option_ids:
          type: array
          uniqueItems: true
          maxItems: 8
          items: { type: string, minLength: 1, maxLength: 128 }
    ChoiceScoringPolicy:
      type: object
      additionalProperties: false
      required: [mode, min_points, max_points, speed_bonus, partial_credit]
      properties:
        mode: { type: string, enum: [none, points] }
        min_points: { type: integer, minimum: 0 }
        max_points: { type: integer, minimum: 0 }
        speed_bonus: { type: boolean }
        partial_credit: { type: boolean }
    ActivityTimingPolicy:
      type: object
      additionalProperties: false
      required: [duration_seconds]
      properties:
        duration_seconds: { type: integer, minimum: 5, maximum: 1200 }
    ActivityResultPolicy:
      type: object
      additionalProperties: false
      required: [show_overall_leaderboard_after]
      properties:
        show_overall_leaderboard_after: { type: boolean }
    ContentSlideContent:
      type: object
      additionalProperties: false
      minProperties: 1
      description: At least one of title, text, or image_url must be non-empty.
      properties:
        title: { type: string, maxLength: 120 }
        text: { type: string, maxLength: 600 }
        image_url: { type: string, maxLength: 4096 }
        image_asset_id: { type: string, format: uuid }
        image_width: { type: integer, minimum: 1, maximum: 8192 }
        image_height: { type: integer, minimum: 1, maximum: 8192 }
        image_alt_text: { type: string, maxLength: 300 }
        image_focal_x: { type: number, format: double, minimum: 0, maximum: 1 }
        image_focal_y: { type: number, format: double, minimum: 0, maximum: 1 }
...
    Slide:
      type: object
      required: [id, revision, position, kind, content]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        revision: { type: integer, format: int64, minimum: 1 }
        position: { type: integer, minimum: 0 }
        kind: { type: string, enum: [activity, content] }
        content: { $ref: '#/components/schemas/EditableSlideContent' }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Optional safe human-readable detail; clients should branch on error, not this text.
        field_errors:
          type: object
          description: Optional field-scoped machine-readable validation codes.
          additionalProperties:
            type: array
            items: { type: string }
        retry_after_seconds:
          type: integer
          minimum: 0
          description: Optional retry delay for bounded/rate-limited operations.
        request_id:
          type: string
          description: Optional correlation identifier safe to expose to the client.
...
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: No valid current session exists.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    CSRFError:
      description: The CSRF header and cookie do not match a valid current session.
      content:
        application/json:
              schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: The requested resource does not exist or is not visible to the current account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    EditConflict:
      description: The resource changed after the client loaded it; reload the latest representation before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }